Add files via upload
First, boosted regression tree (BRT) models of Canada Warbler abundance in Alberta are run. Then, Landis scenarios are run and future abundance of Canada Warbler is predicted from Landis outputs with the BRT models. Current and future predicted abundance of Canada Warbler and known occurrences of Canada Warbler in Alberta are used to run conservation planning exercises in Zonation outside of R. Then, the outputs of those Zonation exercises are used to generate custom maps and runtime plots in this folder showing: 1) the places to prioritize for conservation and management as Canada Warbler habitat, and 2) how much habitat you need to maintain a certain % of the current population under different situations.